ocr_testing/ocr_engines/requirements.txt
easyocr>=1.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.20.0
opencv-python-headless>=4.5.0
argparse>=1.4.0
nltk>=3.6.0
matplotlib>=3.3.0 